Rename llm-config-normalize to azure-openai
