+++ src/lib/azure-openai.ts
